Development/GenAI/ImmigrationChatbot/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "immigration-chatbot"
version = "0.1.0"
description = "MVP chatbot for immigration and career uncertainty support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"fastapi>=0.115.0",
	"uvicorn>=0.30.0",
	"sqlalchemy>=2.0.0",
	"psycopg[binary]>=3.2.0",
	"httpx>=0.27.0",
	"google-cloud-firestore>=2.21.0",
]

[project.scripts]
immigration-chatbot = "immigration_chatbot.cli:main"
immigration-chatbot-api = "immigration_chatbot.api:run"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
